hub-core/pyproject.toml
[project]
name = "dombot-hub-core"
version = "0.1.0"
description = "LabOS hub core: metrics, providers, tokens and status aggregation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.0",
    "pydantic>=2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "aiohttp>=3.9",
]

[project.scripts]
hub-core = "hub_core.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hub_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pylint>=4.0.4",
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]

